fix: refactor clientportal service
