+++ backend/core-api/src/modules/clientportal/services/auth/index.ts
@@ -1,21 +1,1 @@
-import {
-  generateTokenPair,
-  setAuthCookie,
-  setAccessTokenCookie,
-  refreshAccessToken,
-  refreshAndSetAuth,
-  clearAuthCookie,
-  createClientPortalToken,
-} from './authService';
-
-export * from './authService';
-
-export const authService = {
-  generateTokenPair,
-  setAuthCookie,
-  setAccessTokenCookie,
-  refreshAccessToken,
-  refreshAndSetAuth,
-  clearAuthCookie,
-  createClientPortalToken,
-};
+export { JwtManager, jwtManager } from './jwtManager';
