+++ backend/core-api/src/modules/clientportal/services/auth/login/index.ts
@@ -1,0 +1,4 @@
+export { loginWithCredentials } from './loginWithCredentials';
+export { loginWithOTP, checkOTPResendLimit } from './loginWithOTP';
+export { sendOTPForLogin } from './sendLoginOTP';
+export { loginWithSocial } from '@/clientportal/services/user/socialAuthService';
