+++ backend/core-api/src/modules/clientportal/services/auth/password/forgot-password.ts
@@ -2,11 +2,10 @@ import { IClientPortalDocument } from '@/clientportal/types/clientPortal';
 import { ICPUserDocument } from '@/clientportal/types/cpUser';
 import * as crypto from 'crypto';
 import { IModels } from '~/connectionResolvers';
-import { AuthenticationError } from '@/clientportal/services/errorHandler';
-import { detectIdentifierType, validatePassword } from './helpers/validators';
-import { setActionCode, validateActionCode } from './helpers/actionCodeHelper';
-import { sendAndStoreOTP } from './helpers/otpSenderHelper';
+import { setActionCode } from '../../helpers/actionCodeHelper';
+import { sendAndStoreOTP } from '../../helpers/otpSenderHelper';
 import { notificationService } from '@/clientportal/services/notification';
+import { detectIdentifierType } from '../../helpers/validators';
 
 const RESET_TOKEN_EXPIRY_HOURS = 1;
 
@@ -105,70 +104,4 @@ export async function forgotPassword(
       models,
     );
   }
-}
-
-function validateResetToken(
-  user: ICPUserDocument | null,
-  code?: string | number,
-): void {
-  if (!user || !user.actionCode) {
-    throw new AuthenticationError('Invalid or expired reset token');
-  }
-  const codeToValidate = code !== undefined ? code : user.actionCode.code;
-  validateActionCode(user, codeToValidate, 'PASSWORD_RESET');
-}
-
-async function updateUserPassword(
-  userId: string,
-  hashedPassword: string,
-  models: IModels,
-): Promise<void> {
-  await models.CPUser.updateOne(
-    { _id: userId },
-    {
-      $set: { password: hashedPassword },
-      $unset: { actionCode: '' },
-    },
-  );
-}
-
-export async function resetPasswordWithToken(
-  token: string,
-  newPassword: string,
-  models: IModels,
-): Promise<ICPUserDocument> {
-  const user = await models.CPUser.findOne({ 'actionCode.code': token });
-  validateResetToken(user, token);
-  validatePassword(newPassword);
-
-  const hashedPassword = await models.CPUser.generatePassword(newPassword);
-  await updateUserPassword(user!._id, hashedPassword, models);
-  return user!;
-}
-
-export async function resetPasswordWithCode(
-  identifier: string,
-  code: string,
-  newPassword: string,
-  clientPortalId: string,
-  models: IModels,
-): Promise<ICPUserDocument> {
-  const identifierType = detectIdentifierType(identifier);
-  const user = await models.CPUser.findByIdentifier(
-    identifier,
-    identifierType,
-    clientPortalId,
-  );
-
-  if (!user || !user.actionCode) {
-    throw new AuthenticationError('Invalid or expired reset token');
-  }
-
-  validateResetToken(user, code);
-  validatePassword(newPassword);
-
-  const hashedPassword = await models.CPUser.generatePassword(newPassword);
-  await updateUserPassword(user._id, hashedPassword, models);
-
-  return user;
 }
