+++ backend/core-api/src/modules/clientportal/services/auth/password/index.ts
@@ -1,0 +1,2 @@
+export * from './forgot-password';
+export * from './reset-password';
