+++ backend/core-api/src/modules/clientportal/services/helpers/codeGenerator.ts
