+++ backend/core-api/src/modules/clientportal/services/helpers/validators.ts
