+++ backend/core-api/src/modules/clientportal/services/index.ts
@@ -1,16 +1,4 @@
-export { authService } from './auth';
+export { jwtManager } from './auth';
 export { cpUserService, contactService, socialAuthService } from './user';
-export {
-  verificationService,
-  otpService,
-  passwordService,
-} from './verification';
+export * from './auth/login';
 export { notificationService, firebaseService } from './notification';
-export * from './errorHandler';
-export * from './helpers/userUtils';
-export * from './helpers/queryBuilders';
-export * from './verification/helpers/otpConfigHelper';
-export * from './verification/helpers/validators';
-export * from './verification/helpers/otpSenderHelper';
-export * from './verification/helpers/actionCodeHelper';
-export * from './helpers/socialAuth';
